History Recorder

History Recorder

A HistoryRecorder runs a simulation and records the trajectory. It returns a history record (MDPHistory or POMDPHistory).

hr = HistoryRecorder(max_steps=100)
pomdp = TigerPOMDP()
policy = RandomPolicy(pomdp)

h = simulate(hr, pomdp, policy)

POMDPSimulators.HistoryRecorder — Type.

A simulator that records the history for later examination

The simulation will be terminated when either

a terminal state is reached (as determined by isterminal() or

the discount factor is as small as eps or

max_steps have been executed

Keyword Arguments: - rng: The random number generator for the simulation - capture_exception::Bool: whether to capture an exception and store it in the history, or let it go uncaught, potentially killing the script - show_progress::Bool: show a progress bar for the simulation - eps - max_steps - sizehint::Int: the expected length of the simulation (for preallocation)

Usage (optional arguments in brackets):

hr = HistoryRecorder()
history = simulate(hr, pomdp, policy, [updater [, init_belief [, init_state]]])

source